Revert "Partially revert "Update fingerprint wrapper for android 12""
This commit was included in my last public build, NagasRamazan (on XDA developers) reported unlock delays

His original message:
Unlocking with fingerprint works with delay when sleeping with screen locked. But after waking up the screen with the power button or the home button, the fingerprint reader unlocks very quickly.
The previous version did not have this delay.

This is the only change i made, and fingerprint works well even without this as no one reported problems with the first public build

This reverts commit 4df70f1a2c000f3fa58de651220c8ef316fa64c7.

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>